requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.18
django-allauth==0.42.0
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==23.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-resized==1.0.2
django-smartfields==1.1.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.1.3
oauthlib==3.1.0
Pillow==9.5.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pydantic==1.10.7
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.3.1
typing_extensions==4.5.0
urllib3==1.25.9
whitenoise==5.1.0
zipp==3.15.0
